requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-CORS
marshmallow
python-dotenv
coolname
pyjwt
pyjwt[crypto]